requirements/base.txt
aiogram==3.0.0b7
python-dotenv==1.0.0
openai==0.28
python-pptx==0.6.19
